Implement block and metadata storage
This commit finishes implementing block and metadata storage for the
SSH/SFTP backend. It also makes some secondary changes in the metadata
module and adds some utilities to support this

Specifically, in order to generate identity tags easier, this adds a new
Hasher struct which hashes the data read from or written to an
underlying reader or writer. This is used to implement changes to the
metadata object's 'save' interface which produces an identity tag as a
side effect.